+++ Runtime/Core/Caretaker.cs
@@ -1,16 +1,20 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using AceLand.Library.Disposable;
+using UnityEngine;
 
 namespace AceLand.MementoService.Core
 {
     internal class Caretaker<T> : DisposableObject
     {
+        public Caretaker(int historyLimit) =>
+            this.historyLimit = historyLimit; 
+        
         private Stack<MementoState<T>> _undoStack = new();
         private readonly Stack<MementoState<T>> _redoStack = new();
-        private readonly int maxHistory = Memento.Settings.UndoLimit;
+        private readonly int historyLimit;
         
-        public int UndoCount => _undoStack.Count - 1;
+        public int UndoCount => Mathf.Max(_undoStack.Count - 1, 0);
         public int RedoCount => _redoStack.Count;
 
         protected override void DisposeManagedResources()
@@ -23,7 +27,7 @@ namespace AceLand.MementoService.Core
             _undoStack.Push(memento);
             _redoStack.Clear();
 
-            if (_undoStack.Count <= maxHistory) return;
+            if (_undoStack.Count <= historyLimit) return;
             
             _undoStack = new Stack<MementoState<T>>(_undoStack.Reverse().Skip(1));
         }