+++ Runtime/Core/IMementoService.cs
@@ -11,6 +11,8 @@
     public interface IMementoService<T> : IMementoService
     {
         void SaveState(T state);
+        T Undo();
+        T Redo();
         bool TryUndo(out T state);
         bool TryRedo(out T state);
     }