+++ Runtime/MementoService.cs
@@ -1,4 +1,5 @@
-﻿using AceLand.MementoService.Core;
+﻿using System;
+using AceLand.MementoService.Core;
 using AceLand.MementoService.ProjectSetting;
 using UnityEngine;
 
@@ -8,8 +9,8 @@ namespace AceLand.MementoService
     {
         private static AceLandMementoSettings Settings => Memento.Settings; 
         
-        private readonly Originator<T> _originator = new();
-        private readonly Caretaker<T> _caretaker = new();
+        private readonly Originator<T> _originator;
+        private readonly Caretaker<T> _caretaker;
 
         public int UndoCount => _caretaker.UndoCount;
         public int RedoCount => _caretaker.RedoCount;
@@ -20,6 +21,17 @@ namespace AceLand.MementoService
             _caretaker.AddMementoState(_originator.SaveToMemento());
             
             if (Settings.IsLogLevel) Debug.Log($"Saved Memento State : {typeof(T).Name}");
+        }
+
+        public T Undo()
+        {
+            var mementoState = _caretaker.Undo();
+            if (mementoState == null)
+                throw new Exception($"No recovered Memento state for type {typeof(T).Name}. Please check UndoCount before Undo, or use TryUndo method.");
+            
+            _originator.RestoreFromMemento(mementoState);
+            if (Settings.IsLogLevel) Debug.Log($"Undo Memento State : {typeof(T).Name}");
+            return _originator.GetState();
         }
         
         public bool TryUndo(out T state)
@@ -33,6 +45,17 @@ namespace AceLand.MementoService
             
             if (Settings.IsLogLevel) Debug.Log($"Undo Memento State : {typeof(T).Name}");
             return true;
+        }
+
+        public T Redo()
+        {
+            var mementoState = _caretaker.Redo();
+            if (mementoState == null)
+                throw new Exception($"No recovered Memento state for type {typeof(T).Name}. Please check RedoCount before Redo, or use TryRedo method.");
+            
+            _originator.RestoreFromMemento(mementoState);
+            if (Settings.IsLogLevel) Debug.Log($"Redo Memento State : {typeof(T).Name}");
+            return _originator.GetState();
         }
         
         public bool TryRedo(out T state)