+++ Runtime/ProjectSetting/AceLandMementoSettings.cs
@@ -1,7 +1,6 @@
 ﻿using AceLand.Library.BuildLeveling;
 using AceLand.Library.ProjectSetting;
 using UnityEngine;
-using UnityEngine.Serialization;
 
 namespace AceLand.MementoService.ProjectSetting
 {
@@ -9,11 +8,11 @@ namespace AceLand.MementoService.ProjectSetting
     {
         [Header("Settings")]
         [SerializeField] private MementoServiceMode mementoServiceMode = MementoServiceMode.GlobalAndLocal;
-        [SerializeField, Min(4)] private int undoLimit = 32;
+        [SerializeField, Min(4)] private int historyLimit = 32;
         [SerializeField] private BuildLevel logLevel = BuildLevel.Development;
         
         public MementoServiceMode MementoServiceMode => mementoServiceMode;
-        public int UndoLimit => undoLimit;
+        public int HistoryLimit => historyLimit;
         public bool IsLogLevel => logLevel.IsAcceptedLevel();
     }
 }